requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
array-record==0.5.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.7.0
blis==0.7.11
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.10
cityscapesScripts==2.2.2
cloudpathlib==0.15.1
cmake==3.25.0
coloredlogs==15.0.1
confection==0.1.3
cymem==2.0.8
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.10
etils==1.5.2
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
flatbuffers==23.5.26
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gradio==3.35.2
gradio_client==0.7.0
grpcio==1.57.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
imgaug==0.4.0
iopath==0.1.9
joblib==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.15.0
langcodes==3.3.0
libclang==16.0.6
linkify-it-py==2.0.2
lit==15.0.7
Markdown==3.4.4
markdown-it-py==2.2.0
matplotlib==3.4.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
ml-dtypes==0.2.0
multidict==6.0.4
munkres==1.1.4
murmurhash==1.0.10
mypy-extensions==1.0.0
natsort==8.4.0
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.9.10
pandas==1.5.3
pathspec==0.11.2
pathtools==0.1.2
pathy==0.10.2
Pillow==9.4.0
ply==3.11
portalocker==2.7.0
preshed==3.0.9
promise==2.3
protobuf==3.20.3
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
PyQt5-sip==12.11.0
pyquaternion==0.9.9
python-multipart==0.0.6
referencing==0.31.0
regex==2023.10.3
requests-oauthlib==1.3.1
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.7.3
semantic-version==2.10.0
sentry-sdk==1.22.2
setproctitle==1.3.2
shapely==2.0.2
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorrt==8.5.3.1
termcolor==2.3.0
thinc==8.2.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchstat==0.0.7
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.8.0
uc-micro-py==1.0.2
ultralytics==8.0.215
uvicorn==0.24.0.post1
wandb==0.15.2
wasabi==1.1.2
weasel==0.3.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
yacs==0.1.8
yarl==1.9.3
